requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.2
sqlparse==0.4.2
whitenoise==5.3.0
